+++ src/collective/documentgenerator/setuphandlers.py
@@ -115,6 +115,7 @@ def install_demo(context):
             container=pod_folder,
             exclude_from_nav=True,
             style_template=[style_template.UID()],
+            rename_page_styles=True,
         )
     loop_template = getattr(pod_folder, 'loop_template')
 
@@ -129,7 +130,7 @@ def install_demo(context):
                 filename=u'modele_general.odt',
             ),
             container=pod_folder,
-            exclude_from_nav=True
+            exclude_from_nav=True,
         )
 
     if not hasattr(pod_folder, 'test_template_multiple'):
@@ -154,6 +155,7 @@ def install_demo(context):
                     'do_rendering': True,
                 }
             ],
+            rename_page_styles=False,
         )
 
     if not hasattr(pod_folder, 'test_template_bis'):
@@ -183,7 +185,8 @@ def install_demo(context):
                     'name': 'details',
                     'value': '1',
                 }
-            ]
+            ],
+            rename_page_styles=False,
         )
 
     if not hasattr(pod_folder, 'test_ods_template'):
@@ -201,6 +204,7 @@ def install_demo(context):
             pod_formats=['ods', 'xls', ],
             pod_portal_types=['Document'],
             style_template=[style_template.UID()],
+            rename_page_styles=False,
         )
 
     if not hasattr(pod_folder, 'test_template_possibly_mailed'):
@@ -218,6 +222,7 @@ def install_demo(context):
             pod_formats=['odt'],
             pod_portal_types=['Folder'],
             mailing_loop_template=loop_template.UID(),
+            rename_page_styles=False,
         )